refactor: added and refactored app text field
